create saveWorld function and log time
